+++ Services/Interfaces/INamingService.cs
@@ -6,5 +6,7 @@ namespace AutomatedClashRunner.Services.Interfaces
         string GetClashTestName(string modelDisplayName, string manualSetName);
         string GetToolsTestClashName(string modelDisplayName);
         string GetBaseBuildClashName(string modelDisplayName);
+        string GetConstructabilityClashName(string modelDisplayName);
+        string GetConstructabilityClashName(System.Collections.Generic.List<AutomatedClashRunner.Models.ModelSourceNode> models);
     }
 }